NEW: Log job output into the job messages.
If a job is long running, waiting for an error/completion message may
be insufficient. In a server environment without SSH access, this can
make it hard to diagnose issues.

This change uses output buffering to grab job output 256 chars at a time
and push it up as job messages.

In addition, the list view is refactored to only show the most recent
message, and a “Messages” tab added to the detail view with the full
log.