FIX Remove stalled job if implementation not found (#350)
* Remove stalled job if implementation not found

If a job class is renamed or removed, remove the related stalled jobs rather than throwing an exception (which prevents the rest of the jobs from running).

* Handle case if jobDescriptor is deleted.

When `refreshDescriptor()` updates $this->descriptor from the database, potentially this object can be null. This change checks to ensure `$this->descriptor` is set before calling any methods.